Add config back

Closes #1912

See merge request scclouds/scclouds!771

+++ engine/schema/src/main/java/com/cloud/vm/dao/VmStatsDao.java
@@ -75,8 +75,9 @@ public interface VmStatsDao extends GenericDao<VmStatsVO, Long> {
     /**
      * Removes (expunges) all VM stats with {@code timestamp} less than
      * a given Date.
+     *
      * @param limit the maximum date to keep stored. Records that exceed this limit will be removed.
      */
-    Integer removeAllByTimestampLessThan(Date limit, Long limitPerQuery);
+    long removeAllByTimestampLessThan(Date limit, Long limitPerQuery);
 
 }
